Fix I_GetFreeMem

+++ src/dummy/i_system.c
@@ -8,7 +8,7 @@ UINT8 graphics_started = 0;
 
 UINT8 keyboard_started = 0;
 
-UINT32 I_GetFreeMem(UINT32 *total)
+size_t I_GetFreeMem(UINT32 *total)
 {
 	*total = 0;
 	return 0;
